Make template directory configurable on the command line
The new --template-dir= flag allows the user to optionally specify the
directory where the templates are located. It will default to the old way
of using $GOPATH to find the template directory in the source directory of
startpage. Being able to use --template-dir= can be especially useful, if
you want to launch startpage when $GOPATH is not (yet) defined (e.g. when
using systemd to bring up startpage).